Refactor movement methods and update mixins
- Replaced `stopMove()` with `stop()` in various mixins for consistency.
- Updated diagonal speed calculation in `MovementByJoystick` mixin.
- Changed method names from `moveFromAngle` to `moveByAngle` for clarity.
- Removed the `SimpleMovement` mixin as it was deemed unnecessary.
- Adjusted enemy and ally classes to use the new movement methods.
- Updated extensions to reflect changes in direction handling and movement methods.
- Ensured that NPCs and players utilize the updated movement logic.
