+++ lib/bonfire.dart
@@ -7,12 +7,12 @@ export 'package:bonfire/base/bonfire_game_interface.dart';
 export 'package:bonfire/base/game_component.dart';
 export 'package:bonfire/behavior/behavior.dart';
 export 'package:bonfire/camera/camera_config.dart';
-export 'package:bonfire/collision/block_movement_collision.dart';
+export 'package:bonfire/collision/simple_collision.dart';
 export 'package:bonfire/color_filter/game_color_filter.dart';
 export 'package:bonfire/decoration/decoration.dart';
 export 'package:bonfire/decoration/decoration_with_collision.dart';
 export 'package:bonfire/forces/forces_2d.dart';
-export 'package:bonfire/forces/handle_forces.dart';
+export 'package:bonfire/forces/forces.dart';
 export 'package:bonfire/game_interface/game_interface.dart';
 export 'package:bonfire/game_interface/interface_component.dart';
 export 'package:bonfire/game_interface/text_interface_component.dart';
@@ -37,7 +37,6 @@ export 'package:bonfire/map/tiled/world_map_by_tiled.dart';
 export 'package:bonfire/map/util/world_map_reader.dart';
 export 'package:bonfire/map/world_map.dart';
 export 'package:bonfire/mixins/attackable.dart';
-export 'package:bonfire/mixins/elastic_collision.dart';
 export 'package:bonfire/mixins/flip_render.dart';
 export 'package:bonfire/mixins/follower.dart';
 export 'package:bonfire/mixins/interval_checker.dart';
@@ -52,6 +51,7 @@ export 'package:bonfire/mixins/pushable.dart';
 export 'package:bonfire/mixins/random_movement.dart';
 export 'package:bonfire/mixins/sensor.dart';
 export 'package:bonfire/mixins/shader/use_shader.dart';
+export 'package:bonfire/collision/simple_elastic_collision.dart';
 export 'package:bonfire/mixins/tile_recognizer.dart';
 export 'package:bonfire/mixins/update_camera_by_pinch_gesture.dart';
 export 'package:bonfire/mixins/use_assets_loader.dart';
