+++ lib/mixins/jumper.dart
@@ -16,7 +16,7 @@ enum JumpingStateEnum {
 ///     GravityForce2D(),
 ///   ],
 /// )
-mixin Jumper on Movement, BlockMovementCollision {
+mixin Jumper on Movement, SimpleCollision {
   final double _defaultJumpSpeed = 150;
   bool isJumping = false;
   JumpingStateEnum jumpingState = JumpingStateEnum.idle;
@@ -44,17 +44,14 @@ mixin Jumper on Movement, BlockMovementCollision {
   }
 
   @override
-  void onBlockedMovement(
-    PositionComponent other,
-    CollisionData collisionData,
-  ) {
+  void onMovementBlocked(PositionComponent other, CollisionData collisionData) {
     if (isJumping &&
-        lastDirectionVertical.isDownSide &&
+        direction.isDownSide &&
         collisionData.direction.isDownSide) {
       _currentJumps = 0;
       isJumping = false;
     }
-    super.onBlockedMovement(other, collisionData);
+    super.onMovementBlocked(other, collisionData);
   }
 
   @override
@@ -81,16 +78,18 @@ mixin Jumper on Movement, BlockMovementCollision {
 
   @override
   void update(double dt) {
-    if (checkInterval(
-          _tileCollisionCountKey,
-          100,
-          dt,
-          firstCheckIsTrue: false,
-        ) &&
-        !isJumping &&
-        _tileCollisionCount == 0 &&
-        displacement.y.abs() > 0.2) {
-      isJumping = true;
+    final tick = checkInterval(
+      _tileCollisionCountKey,
+      100,
+      dt,
+      firstCheckIsTrue: false,
+    );
+    if (tick) {
+      if (!isJumping &&
+          _tileCollisionCount == 0 &&
+          velocity.y.abs() > 0.2) {
+        isJumping = true;
+      }
     }
     _notifyJump();
     super.update(dt);
@@ -99,7 +98,7 @@ mixin Jumper on Movement, BlockMovementCollision {
   void _notifyJump() {
     JumpingStateEnum newDirection;
     if (isJumping) {
-      if (lastDirectionVertical == Direction.down) {
+      if (direction.isDownSide) {
         newDirection = JumpingStateEnum.down;
       } else {
         newDirection = JumpingStateEnum.up;
@@ -114,9 +113,9 @@ mixin Jumper on Movement, BlockMovementCollision {
   }
 
   @override
-  void stopMove({bool forceIdle = false, bool isX = true, bool isY = true}) {
+  void stop() {
     if (!isJumping) {
-      super.stopMove(forceIdle: forceIdle, isX: isX, isY: isY);
+      super.stop();
     }
   }
 }
