+++ lib/mixins/mixins.dart
@@ -1,4 +1,3 @@
-export 'simple_collision.dart';
-export 'simple_elastic_collision.dart';
-export 'simple_forces.dart';
-export 'simple_movement.dart';
+export '../collision/simple_collision.dart';
+export '../collision/simple_elastic_collision.dart';
+export '../forces/forces.dart';
