+++ lib/npc/ally/rotation_ally.dart
@@ -40,12 +40,12 @@ class RotationAlly extends Ally with UseSpriteAnimation, UseAssetsLoader {
     );
   }
 
-  @override
-  void moveFromAngle(double angle, {double? speed}) {
-    setAnimation(animRun);
-    this.angle = angle;
-    super.moveFromAngle(angle, speed: speed);
-  }
+  // @override
+  // void moveFromAngle(double angle, {double? speed}) {
+  //   setAnimation(animRun);
+  //   this.angle = angle;
+  //   super.moveFromAngle(angle, speed: speed);
+  // }
 
   @override
   void idle() {
