+++ lib/npc/ally/simple_ally.dart
@@ -27,8 +27,7 @@ class SimpleAlly extends Ally with DirectionAnimation {
     super.receivesAttackFrom,
   }) {
     this.animation = animation;
-    lastDirection = initDirection;
-    lastDirectionHorizontal =
-        initDirection == Direction.left ? Direction.left : Direction.right;
+    direction = initDirection;
+
   }
 }
