+++ lib/npc/enemy/rotation_enemy.dart
@@ -28,12 +28,12 @@ class RotationEnemy extends Enemy with UseSpriteAnimation, UseAssetsLoader {
     );
   }
 
-  @override
-  void moveFromAngle(double angle, {double? speed}) {
-    setAnimation(animRun);
-    this.angle = angle;
-    super.moveFromAngle(angle, speed: speed);
-  }
+  // @override
+  // void moveFromAngle(double angle, {double? speed}) {
+  //   setAnimation(animRun);
+  //   this.angle = angle;
+  //   super.moveFromAngle(angle, speed: speed);
+  // }
 
   @override
   void idle() {
