+++ lib/npc/enemy/simple_enemy.dart
@@ -15,8 +15,6 @@ class SimpleEnemy extends Enemy with DirectionAnimation {
     super.receivesAttackFrom,
   }) {
     this.animation = animation;
-    lastDirection = initDirection;
-    lastDirectionHorizontal =
-        initDirection == Direction.left ? Direction.left : Direction.right;
+    direction = initDirection;
   }
 }
