+++ lib/npc/rotation_npc.dart
@@ -38,12 +38,12 @@ class RotationNpc extends Npc with UseSpriteAnimation, UseAssetsLoader {
     );
   }
 
-  @override
-  void moveFromAngle(double angle, {double? speed}) {
-    setAnimation(animRun);
-    this.angle = angle;
-    super.moveFromAngle(angle, speed: speed);
-  }
+  // @override
+  // void moveFromAngle(double angle, {double? speed}) {
+  //   setAnimation(animRun);
+  //   this.angle = angle;
+  //   super.moveFromAngle(angle, speed: speed);
+  // }
 
   @override
   void idle() {
