+++ lib/scene_builder/actions/move_while_scene_action.dart
@@ -18,7 +18,7 @@ class MoveWhileSceneAction extends SceneAction {
       doThis(component);
       return false;
     } else {
-      component.stopMove();
+      component.stop();
       return true;
     }
   }
